requirements.txt
alembic==1.7.4
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
greenlet==1.1.2
psycopg2-binary==2.9.1
PySocks==1.7.1
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.25
tornado==6.1
tzdata==2021.1
tzlocal==3.0
wcwidth==0.2.5
